feat: move faq into redesign
