+++ frontend-v2/src/components/navigation/TopNav.tsx
@@ -1,6 +1,5 @@
 import {
     useAuth,
-    isAdministrator,
     canSwitchFacility,
     handleLogout,
     isFacilityAdmin
@@ -13,15 +12,13 @@ import {
     DropdownMenuTrigger
 } from '@/components/ui/dropdown-menu';
 import { Button } from '@/components/ui/button';
-import { Building2, LogOut, CircleHelp } from 'lucide-react';
+import { Building2, LogOut } from 'lucide-react';
 import { usePageTitle } from '@/contexts/PageTitleContext';
 
 export default function TopNav({
-    facilities,
-    onToggleHelpCenter
+    facilities
 }: {
     facilities?: Facility[];
-    onToggleHelpCenter: () => void;
 }) {
     const { user } = useAuth();
     const { pageTitle } = usePageTitle();
@@ -90,15 +87,6 @@ export default function TopNav({
                     </DropdownMenu>
                 ) : null}
 
-                {!isAdministrator(user) && (
-                    <Button
-                        variant="ghost"
-                        size="sm"
-                        onClick={onToggleHelpCenter}
-                    >
-                        <CircleHelp className="h-4 w-4" />
-                    </Button>
-                )}
 
                 <DropdownMenu>
                     <DropdownMenuTrigger asChild>
